backend/pyproject.toml
[project]
name = "chatddx-backend"
version = "0.0.0+dev"
description = "chatddx backend"
requires-python = ">=3.12"
dependencies = [
  "Django>=6.0.2",
  "celery>=5.6.2",
  "dj-user-login-history>=1.0.6",
  "django-celery-results>=2.6.0",
  "django-modeltranslation>=0.19.19",
  "django-ninja>=1.6.2",
  "django-typer>=3.7.0",
  "django-unfold>=0.87.0",
  "flower>=2.0.1",
  "gunicorn>=25.0.1",
  "jsonschema>=4.26.0",
  "markdown2>=2.5.4",
  "openai>=2.16.0",
  "psycopg[binary]>=3.3.2",
  "pydantic-ai>=1.55.0",
  "redis>=7.1.0",
  "time-machine>=3.2.0",
  "tomli>=2.4.0",
]

[dependency-groups]
dev = [
  { include-group = "lint" },
  { include-group = "test" },
  { include-group = "typing" },
]
lint = ["isort>=8.0.0", "ruff>=0.15.0"]
test = [
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "pytest-django>=4.11.1",
  "pytest-recording>=0.13.4",
]
typing = ["django-types>=0.23.0", "pyright>=1.1.408"]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.django-stubs]
django_settings_module = "chatddx_backend.settings"

[tool.pyright]
include = ["src"]
ignore = ["**/migrations/*"]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
